python_backend/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=5.4.1

# Exchange connectivity
aiohttp>=3.8.0
websockets>=10.0
ccxt>=2.5.0  # Optional: for additional exchange support

# API framework
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
pydantic>=1.9.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
alembic>=1.8.0

# Redis
redis>=4.3.0
aioredis>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Monitoring
prometheus-client>=0.14.0
opentelemetry-api>=1.12.0
opentelemetry-sdk>=1.12.0
opentelemetry-instrumentation-fastapi>=0.33b0

# Alternative data sources
tweepy>=4.10.0  # Twitter API
praw>=7.6.0  # Reddit API
newsapi-python>=0.2.6  # News API

# Utilities
python-dotenv>=0.19.0
click>=8.0.0
rich>=12.5.0  # Better console output
apscheduler>=3.10.0  # Task scheduling for automatic updates

# Testing
pytest>=7.1.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0

# Development
black>=22.6.0
flake8>=4.0.0
mypy>=0.961
pre-commit>=2.20.0